exception catching;
restaured contract id code

+++ hedge-fund-core/src/main/java/com/onenow/portfolio/BrokerController.java
@@ -925,7 +925,7 @@ public class BrokerController implements EWrapper {
     @Override public void realtimeBar(int reqId, long time, double open, double high, double low, double close, long volume, double wap, int count) {
     	IRealTimeBarHandler handler = m_realTimeBarMap.get( reqId);
 		if (handler != null) {
-			EventActivityHistory bar = new EventActivityHistory(reqId, time, high, low, open, close, wap, (int) volume, count);
+			EventActivityHistory bar = new EventActivityHistory(reqId, time, high, low, open, close, wap, volume, count);
 			handler.realtimeBar( bar);
 		}
 		recEOM();
